Backend/requirements.txt
Flask
pandas
PyPDF2
pytesseract
Pillow
google-generativeai
xlsxwriter
